change uint32_t to uint_t. Same with int32_t
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/vplants/vplants/trunk/PlantGL@4276 ab253cce-29fb-0310-bb2f-979600cdbdeb

+++ src/cpp/plantgl/scenegraph/core/sceneobject.cpp
@@ -76,8 +76,8 @@ void SceneObject::setName( const std::string& name ) {
   __name = name;
 }
 
-uint32_t SceneObject::getId( ) const {
-  return (uint32_t)this;
+size_t SceneObject::getId( ) const {
+  return (size_t)this;
 }
 
 const std::string&
