+++ src/cpp/plantgl/scenegraph/core/sceneobject.h
@@ -151,7 +151,7 @@ public:
   virtual bool apply( Action& action ) = 0;
 
   /// Returns a unique id to identify \e self. 
-  virtual uint32_t getId( ) const ;
+  virtual size_t getId( ) const ;
 
   /// Returns the name of \e self.
   const std::string& getName( ) const ;
